Fix typescript codgen and update integration tests.
