+++ ci/integration-tests/src/helpers/helper.rs
@@ -61,7 +61,7 @@ pub fn create_dao(
             label: "DAO DAO Voting Module".to_string(),
         },
         proposal_modules_instantiate_info: vec![ModuleInstantiateInfo {
-            code_id: chain.orc.contract_map.code_id("cw_proposal_single")?,
+            code_id: chain.orc.contract_map.code_id("cwd_proposal_single")?,
             msg: to_binary(&cwd_proposal_single::msg::InstantiateMsg {
                 min_voting_period: None,
                 threshold: Threshold::ThresholdQuorum {
@@ -74,7 +74,7 @@ pub fn create_dao(
                 close_proposal_on_execution_failure: false,
                 pre_propose_info: PreProposeInfo::ModuleMayPropose {
                     info: ModuleInstantiateInfo {
-                        code_id: chain.orc.contract_map.code_id("cw_pre_propose_single")?,
+                        code_id: chain.orc.contract_map.code_id("cwd_pre_propose_single")?,
                         msg: to_binary(&cwd_pre_propose_single::InstantiateMsg {
                             deposit_info: Some(UncheckedDepositInfo {
                                 denom: DepositToken::VotingModuleToken {},
@@ -98,14 +98,14 @@ pub fn create_dao(
 
     chain
         .orc
-        .instantiate("cw_core", op_name, &msg, key, None, vec![])?;
+        .instantiate("cwd_core", op_name, &msg, key, None, vec![])?;
 
     let res = chain
         .orc
-        .query("cw_core", &cwd_core::msg::QueryMsg::DumpState {})?;
+        .query("cwd_core", &cwd_core::msg::QueryMsg::DumpState {})?;
 
     Ok(DaoState {
-        addr: chain.orc.contract_map.address("cw_core")?,
+        addr: chain.orc.contract_map.address("cwd_core")?,
         state: res.data()?,
     })
 }
