--- typescript/contracts/cw-pre-propose-base-proposal-single/CwPreProposeBaseProposalSingleContract.ts
@@ -1,255 +1,0 @@
-/**
-* This file was automatically generated by @cosmwasm/ts-codegen@0.5.8.
-* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
-* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
-*/
-
-import { CosmWasmClient, ExecuteResult, SigningCosmWasmClient } from "@cosmjs/cosmwasm-stargate";
-import { StdFee } from "@cosmjs/amino";
-export type ExecuteMsgForProposeMessageAnd_Empty = {
-  propose: {
-    msg: ProposeMessage;
-    [k: string]: unknown;
-  };
-} | {
-  update_config: {
-    deposit_info?: UncheckedDepositInfo | null;
-    open_proposal_submission: boolean;
-    [k: string]: unknown;
-  };
-} | {
-  withdraw: {
-    denom?: UncheckedDenom | null;
-    [k: string]: unknown;
-  };
-} | {
-  extension: {
-    msg: Empty;
-    [k: string]: unknown;
-  };
-} | {
-  proposal_hook: ProposalHookMsg;
-};
-export type ProposeMessage = {
-  propose: {
-    description: string;
-    msgs: CosmosMsgForEmpty[];
-    title: string;
-    [k: string]: unknown;
-  };
-};
-export type CosmosMsgForEmpty = {
-  bank: BankMsg;
-} | {
-  custom: Empty;
-} | {
-  staking: StakingMsg;
-} | {
-  distribution: DistributionMsg;
-} | {
-  stargate: {
-    type_url: string;
-    value: Binary;
-    [k: string]: unknown;
-  };
-} | {
-  ibc: IbcMsg;
-} | {
-  wasm: WasmMsg;
-} | {
-  gov: GovMsg;
-};
-export type BankMsg = {
-  send: {
-    amount: Coin[];
-    to_address: string;
-    [k: string]: unknown;
-  };
-} | {
-  burn: {
-    amount: Coin[];
-    [k: string]: unknown;
-  };
-};
-export type Uint128 = string;
-export type StakingMsg = {
-  delegate: {
-    amount: Coin;
-    validator: string;
-    [k: string]: unknown;
-  };
-} | {
-  undelegate: {
-    amount: Coin;
-    validator: string;
-    [k: string]: unknown;
-  };
-} | {
-  redelegate: {
-    amount: Coin;
-    dst_validator: string;
-    src_validator: string;
-    [k: string]: unknown;
-  };
-};
-export type DistributionMsg = {
-  set_withdraw_address: {
-    address: string;
-    [k: string]: unknown;
-  };
-} | {
-  withdraw_delegator_reward: {
-    validator: string;
-    [k: string]: unknown;
-  };
-};
-export type Binary = string;
-export type IbcMsg = {
-  transfer: {
-    amount: Coin;
-    channel_id: string;
-    timeout: IbcTimeout;
-    to_address: string;
-    [k: string]: unknown;
-  };
-} | {
-  send_packet: {
-    channel_id: string;
-    data: Binary;
-    timeout: IbcTimeout;
-    [k: string]: unknown;
-  };
-} | {
-  close_channel: {
-    channel_id: string;
-    [k: string]: unknown;
-  };
-};
-export type Timestamp = Uint64;
-export type Uint64 = string;
-export type WasmMsg = {
-  execute: {
-    contract_addr: string;
-    funds: Coin[];
-    msg: Binary;
-    [k: string]: unknown;
-  };
-} | {
-  instantiate: {
-    admin?: string | null;
-    code_id: number;
-    funds: Coin[];
-    label: string;
-    msg: Binary;
-    [k: string]: unknown;
-  };
-} | {
-  migrate: {
-    contract_addr: string;
-    msg: Binary;
-    new_code_id: number;
-    [k: string]: unknown;
-  };
-} | {
-  update_admin: {
-    admin: string;
-    contract_addr: string;
-    [k: string]: unknown;
-  };
-} | {
-  clear_admin: {
-    contract_addr: string;
-    [k: string]: unknown;
-  };
-};
-export type GovMsg = {
-  vote: {
-    proposal_id: number;
-    vote: VoteOption;
-    [k: string]: unknown;
-  };
-};
-export type VoteOption = "yes" | "no" | "abstain" | "no_with_veto";
-export type DepositToken = {
-  token: {
-    denom: UncheckedDenom;
-    [k: string]: unknown;
-  };
-} | {
-  voting_module_token: {
-    [k: string]: unknown;
-  };
-};
-export type UncheckedDenom = {
-  native: string;
-} | {
-  cw20: string;
-};
-export type DepositRefundPolicy = "always" | "only_passed" | "never";
-export type ProposalHookMsg = {
-  new_proposal: {
-    id: number;
-    proposer: string;
-    [k: string]: unknown;
-  };
-} | {
-  proposal_status_changed: {
-    id: number;
-    new_status: string;
-    old_status: string;
-    [k: string]: unknown;
-  };
-};
-export interface Coin {
-  amount: Uint128;
-  denom: string;
-  [k: string]: unknown;
-}
-export interface Empty {
-  [k: string]: unknown;
-}
-export interface IbcTimeout {
-  block?: IbcTimeoutBlock | null;
-  timestamp?: Timestamp | null;
-  [k: string]: unknown;
-}
-export interface IbcTimeoutBlock {
-  height: number;
-  revision: number;
-  [k: string]: unknown;
-}
-export interface UncheckedDepositInfo {
-  amount: Uint128;
-  denom: DepositToken;
-  refund_policy: DepositRefundPolicy;
-  [k: string]: unknown;
-}
-export interface InstantiateMsgForEmpty {
-  deposit_info?: UncheckedDepositInfo | null;
-  extension: Empty;
-  open_proposal_submission: boolean;
-  [k: string]: unknown;
-}
-export type QueryMsgForEmpty = {
-  proposal_module: {
-    [k: string]: unknown;
-  };
-} | {
-  dao: {
-    [k: string]: unknown;
-  };
-} | {
-  config: {
-    [k: string]: unknown;
-  };
-} | {
-  deposit_info: {
-    proposal_id: number;
-    [k: string]: unknown;
-  };
-} | {
-  extension: {
-    msg: Empty;
-    [k: string]: unknown;
-  };
-};
