+++ typescript/packages/cwd-macros/CwdMacrosContract.ts
